creative_autogen/requirements.txt
openai>=1.0.0
beautifulsoup4
pillow
pinecone-client
google-cloud-vision
requests
pyyaml
